+++ CHANGES.md
@@ -11,6 +11,11 @@
 
 ## develop
 
+- [UPDATE] WebRTC のバージョンを M88 (4324@{#3}) に上げる
+    - @melpon @torikizi
+- [UPDATE] Boost のバージョンを 1.75.0 に上げる
+- [UPDATE] nlohmann/json を Boost.JSON に変更
+
 ## 2020.10
 
 - [UPDATE] WebRTC のバージョンを M87 (4280@{#10}) に上げる